feat: add Navbar component and refactor WeatherHome to use it; update background color in styles
