+++ frontend/tailwind.config.js
@@ -11,8 +11,7 @@ export default {
         secondary2: "#B4B4B4",
         secondary3: "#4D4D4D",
         secondary4: "#FAFAFA",
-        background: "#171717",
-        background2: "#1F1F1F",
+        background: "#f8f8f8",
         border: "#292929",
         borderHover: "#454545",
         delete: "#dc2626",
